requirements.txt
alembic==0.8.8
amqp==1.4.9
aniso8601==1.2.0
anyjson==0.3.3
bcrypt==3.1.1
billiard==3.3.0.23
blinker==1.4
celery==3.1.25
cffi==1.8.3
click==6.6
colorama==0.3.7
cssselect==1.0.0
decorator==4.0.10
Flask==0.11.1
Flask-Bcrypt==0.7.1
Flask-Celery-Helper==1.1.0
Flask-Login==0.4.0
Flask-Migrate==2.0.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.13.1
ipython==5.1.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.37
lxml==3.6.0
Mako==1.0.4
MarkupSafe==0.23
pickleshare==0.7.4
prompt-toolkit==1.0.8
pycparser==2.17
Pygments==2.1.3
python-dateutil==2.6.0
python-editor==1.0.1
pytz==2016.7
redis==2.10.5
requests==2.12.3
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.2
traitlets==4.3.1
vine==1.1.3
wcwidth==0.1.7
Werkzeug==0.11.11
win-unicode-console==0.5
WTForms==2.1
